Moved each communication to own files
